Big Five Personality Test

This data was collected in 2012 through an interactive online personality test. This self-report test measures the big five personality traits using the IPIP Big-Five Factor Markers. The items were rated on a five point scale where 1=Disagree, 3=Neutral, 5=Agree and 0=Missed. The big five personality traits are the best accepted and most commonly used model of personality in academic psychology. If you take a college course in personality psychology, this is what you will learn about. The big five come from the statistical study of responses to personality items. Using a technique called factor analysis researchers can look at the responses of people to hundreds of personality items and ask the question what is the best was to summarize an individual?. This has been done with many samples from all over the world and the general result is that, while there seem to be unlimited personality variables, five stand out from the pack in terms of explaining a lot of a persons answers to questions about their personality: extraversion, neuroticism, agreeableness, conscientiousness and openness to experience. The big-five are not associated with any particular test, a variety of measures have been developed to measure them. This test uses the Big-Five Factor Markers from the International Personality Item Pool, developed by Goldberg (1992).

big_five

Format

A data.frame with 200 rows and 7 columns

Age

Age in years

Gender

f = female, m = male

Extraversion

Average degree of extraversion measured with 10 items (outgoing/energetic vs. solitary/reserved).

Neuroticism

Average degree of neuroticism with 10 items (sensitive/nervous vs. secure/confident).

Agreeableness

Average degree of agreeableness measured with 10 items (friendly/compassionate vs. challenging/detached).

Conscientiousness

Average degree of conscientiousness measured with 10 items (efficient/organized vs. easy-going/careless).

Openness

Average degree of opennes to experience measured with 10 items (inventive/curious vs. consistent/cautious).

Source

https://openpsychometrics.org/_rawdata/ (modified)